Reorganized test files.
